#20 #23 searchbar and navlinks plus file restructure
