+++ src/App.js
@@ -1,13 +1,11 @@
 import React from "react";
-import Routes from "./Containers/Routes";
+import Routes from "./routes/Routes";
 import "./style.css";
 
 export default function App() {
     return (
         <div className="App">
-            <center>
-                <Routes />
-            </center>
+            <Routes />
         </div>
     );
 }
