--- src/Navbar/nav.css
@@ -1,15 +1,0 @@
-.main_div{
-    display: flex;
-    width:100%;
-    height:4rem;
-    background-color: #e4e0d2;
-    overflow: hidden;
-}
-
-
-
-
-
-
-
-
