+++ src/components/Navbar/nav.css
@@ -1,0 +1,8 @@
+.main_div {
+    display: flex;
+    width: 100%;
+    height: 4rem;
+    overflow: hidden;
+    justify-content: end;
+    margin: 1.6rem 0 0 0;
+}
