+++ src/pages/Home/myths.css
@@ -1,3 +1,7 @@
+.title {
+    margin: 23rem 0 0 0;
+}
+
 .Kamala-pic-1 {
     height: 30rem;
     width: 50rem;
